Moved SolverOde and SolverForwarEuler to the API.
So that end-users can see what type of solvers are available and what methods can be called.
