+++ src/solver/solverforwardeuler_p.h
@@ -17,7 +17,8 @@ limitations under the License.
 #pragma once
 
 #include "solverode_p.h"
-#include "solverforwardeuler.h"
+
+#include "libopencor/solverforwardeuler.h"
 
 namespace libOpenCOR {
