+++ src/solver/solverode_p.h
@@ -17,7 +17,8 @@ limitations under the License.
 #pragma once
 
 #include "solver_p.h"
-#include "solverode.h"
+
+#include "libopencor/solverode.h"
 
 namespace libOpenCOR {
